requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
PyMuPDF==1.26.4
python-docx==0.8.11
pillow==10.1.0
numpy==1.26.4
sentence-transformers==2.2.2
open-clip-torch==2.0.0
torch==2.2.3
qdrant-client==1.2.3
aiofiles==23.1.0
whisper @ git+https://github.com/openai/whisper.git
transformers==4.40.0
paddleocr==2.6.1
easyocr==1.6.2
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
